HABI#42 Deprecating XML Seeder
