+++ src/main/java/com/habicus/seeder/xmlloader/data/DataContainers/Container.java
@@ -20,35 +20,19 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.habicus.test.data.DataContainers;
+package com.habicus.seeder.xmlloader.data.DataContainers;
 
-import com.habicus.core.model.Goal;
+import com.habicus.seeder.xmlloader.data.Loader;
 import java.util.List;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
 
-/** Allows a container to hold a list of user and map to standard POJO */
-@XmlRootElement(name = "GoalContainer")
-@XmlAccessorType(XmlAccessType.FIELD)
-public class GoalContainer implements Container {
-
-  public GoalContainer() {}
-
-  @XmlElement(name = "goal")
-  private List<Goal> goals;
-
-  public List<Goal> getGoals() {
-    return goals;
-  }
-
-  public void setGoals(List<Goal> goals) {
-    this.goals = goals;
-  }
-
-  @Override
-  public List<Goal> getAll() {
-    return goals;
-  }
+/**
+ * Every test dao table will inherit Container in an effort to genericize the calls to grab all dao
+ * in the {@link Loader}. This greatly increases the amount of code reuse we can enable when
+ * iterating through and loading all the test dao into the database.
+ *
+ * @param <T>
+ */
+@Deprecated
+public interface Container<T> {
+  public List<T> getAll();
 }
